friends and notFriends route changed to FriendsClass
