+++ server/models/UserClass.js
@@ -80,70 +80,6 @@ class UserClass {
       return new BadRequestError(error.message);
     }
   }
-
-  static async getFriends(user) {
-    const friendList = [];
-
-    const Friends = Parse.Object.extend("Friends");
-    const query1 = new Parse.Query(Friends);
-    const query2 = new Parse.Query(Friends);
-
-    if (!user) {
-      return new BadRequestError("Unauthorized");
-    }
-    try {
-      query1.equalTo("user1Id", user);
-      query2.equalTo("user2Id", user);
-
-      const compoundQuery = Parse.Query.or(query1, query2);
-
-      compoundQuery.include("*");
-
-      const friends = await compoundQuery.find();
-
-      friends.map((item) => {
-        if (item.get("user1Id").id == user.id) {
-          friendList.push(item.get("user2Id"));
-        } else {
-          friendList.push(item.get("user1Id"));
-        }
-      });
-
-      return friendList;
-    } catch (error) {
-      return new BadRequestError(error.message);
-    }
-  }
-
-  static async getNotFriends(user) {
-    try {
-      const friends = await UserClass.getFriends(user);
-
-      const friendIds = [];
-
-      friends.map((friend) => {
-        if (friend) {
-          friendIds.push(friend.id);
-        }
-      });
-
-      const query = new Parse.Query(Parse.User);
-      query.notEqualTo("objectId", user.id);
-      const users = await query.find();
-
-      const usersNotFriends = [];
-
-      users.map((item, index) => {
-        if (!friendIds.includes(item.id)) {
-          usersNotFriends.push(item);
-        }
-      });
-
-      return usersNotFriends;
-    } catch (error) {
-      return new BadRequestError(error.message);
-    }
-  }
 }
 
 module.exports = UserClass;
